Refactoring and separation of concerns using a separate helper.js file
